refactor: 🛠️ tweak dir arc
