--- packages/dataset/src/kanji.d.ts
@@ -1,2 +1,0 @@
-import { type TegakiStroke } from "tegaki";
-export declare const KANJI_DATA_SET: Array<TegakiStroke>;
