+++ packages/shared/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./general";
 export * from "./payload";
+export * from "./stroke";
