+++ packages/shared/src/stroke.ts
