+++ packages/tegaki/src/index.ts
@@ -2,6 +2,5 @@ export * from "@tegaki/dataset";
 export * from "@tegaki/frontend";
 export * from "@tegaki/backend";
 export * from "@tegaki/shared";
-export type * from "./stroke";
 
 // NOTE: @tegaki/cli is not exported because it is not a library.
